+++ packages/core/test_e2e/constants.ts
@@ -1,0 +1,2 @@
+export const API_KEY = process.env.API_KEY!;
+export const OWNER_MNEMONIC = process.env.OWNER_MNEMONIC!;
